+++ runtime/src/chplmemtrack.c
@@ -804,12 +804,12 @@ void* chpl_memtable_next_entry(void* entry) {
   return (void*)(((memTableEntry*)entry)->nextInBucket);
 }
 
-uint64_t chpl_memtable_entry_addr(void* entry) {
-  return (uint64_t)(((memTableEntry*)entry)->memAlloc);
+uintptr_t chpl_memtable_entry_addr(void* entry) {
+  return (uintptr_t)(((memTableEntry*)entry)->memAlloc);
 }
 
-uint64_t chpl_memtable_entry_size(void* entry) {
+size_t chpl_memtable_entry_size(void* entry) {
   memTableEntry* _entry = (memTableEntry*)entry;
-  return (uint64_t)(_entry->size*_entry->number);
+  return (size_t)(_entry->size*_entry->number);
 }
